requirements.txt
flask-mysql
wtforms
passlib
PyMySQL